Testing improvements
