+++ packages/@livestore/common-cf/src/do-rpc/rpc.test.ts
@@ -8,7 +8,8 @@ import {
   RpcSerialization,
   Stream,
 } from '@livestore/utils/effect'
-import { startWranglerDevServerPromise, Vitest } from '@livestore/utils-dev/node-vitest'
+import { startWranglerDevServerPromise } from '@livestore/utils-dev/node'
+import { Vitest } from '@livestore/utils-dev/node-vitest'
 import { beforeAll, expect } from 'vitest'
 import { TestRpcs } from './test-fixtures/rpc-schema.ts'
