+++ packages/@livestore/utils-dev/src/node-vitest/mod.ts
@@ -1,5 +1,3 @@
 import './global.ts'
 
 export * as Vitest from './Vitest.ts'
-export * from './vitest-docker-compose-setup.ts'
-export * from './vitest-wrangler-setup.ts'
