+++ scripts/package.json
@@ -11,6 +11,7 @@
     "@livestore/utils": "workspace:*",
     "@livestore/utils-dev": "workspace:*",
     "@local/tests-integration": "workspace:*",
+    "@local/tests-sync-provider": "workspace:*",
     "@types/bun": "^1.2.21",
     "@types/node": "catalog:"
   }
