+++ scripts/tsconfig.json
@@ -10,6 +10,7 @@
   "references": [
     { "path": "../packages/@livestore/utils" },
     { "path": "../packages/@livestore/utils-dev" },
+    { "path": "../tests/sync-provider" },
     { "path": "../packages/@livestore/common" },
     { "path": "../tests/integration" }
   ]
