+++ tests/integration/src/tests/node-sync/fixtures/vitest-sync-setup.ts
@@ -1,4 +1,4 @@
-import { startWranglerDevServerPromise } from '@livestore/utils-dev/node-vitest'
+import { startWranglerDevServerPromise } from '@livestore/utils-dev/node'
 
 const { port } = await startWranglerDevServerPromise({ cwd: import.meta.dirname })
 process.env.LIVESTORE_SYNC_PORT = port.toString()
